Change CRUD operations for Project, use uniqid
Each div and each item in projects will be given the same uniqid.
The id will associate one to the other. The CRUD methods will create,
read, update, and delete the project by a given id for the project.

populateStorage was made to start with two projects in the beginning.