Add project CRUD functions
getProjectTasks() will be used in order to display every task in a
particular project.

replaceProject() will be used when a user edits a project and clicks
'save'. It changes all values for properties that are different from
the original project object.

deleteProject() will be used when a user clicks the delete button.